Low coupling and high cohesion
