+++ Ejercicio1/Ejercicio1/Matricula/Matrias.cs
@@ -1,0 +1,8 @@
+using Ejercicio1.Abstracciones;
+namespace Ejercicio1.Matricula;
+public class Materia : IMateria
+{
+    public string Codigo { get; set; }
+    public string Nombre { get; set; }
+    public int Creditos { get; set; }
+}
